modificación del codigo src/model/'user.py' y elimino el creado por error en la raiz
