+++ src/models/user.py
@@ -2,6 +2,12 @@ import uuid
 from sqlmodel import SQLModel, Field
 from typing import Optional
 from src.security import create_jwt_token
+from datetime import datetime
+from enum import Enum
+
+class Role(str, Enum):
+    client = "client"
+    admin = "admin"
 
 class User(SQLModel, table=True):
     id: Optional[str] = Field(
@@ -12,6 +18,8 @@ class User(SQLModel, table=True):
     name: str
     email: str = Field(index=True, unique=True)
     password: str
+    role: Role = Field(default=Role.client)  
+    created_at: datetime = Field(default_factory=lambda: datetime.now(datetime.timezone.utc))
     
 
     def get_jwt_token(self):
@@ -21,4 +29,8 @@ class User(SQLModel, table=True):
             'email': self.email,
         })
         
-    
+class UserSession(SQLModel, table=True):
+    id: Optional[int] = Field(default=None, primary_key=True)
+    user_id: int = Field(foreign_key="user.id")  
+    start_time: datetime = Field(default_factory=lambda: datetime.now(datetime.timezone.utc))
+    end_time: Optional[datetime] = None
